Generate `BatteryLife` with dc/ac/charging times
Uses the data from the `temps` table to generate this.
